fix: some bugs with parallel upload
